Updated templates to use Tookit support for Master Details.
