+++ templates/Pages/MasterDetail.Prism/.template.config/template.json
@@ -18,32 +18,17 @@
     "wts.genGroup": "0",
     "wts.rightClickEnabled":"true" 
   },
-  "sourceName": "MasterDetail",
+  "sourceName": "MasterDetailView",
   "preferNameDirectory": true,
   "PrimaryOutputs": [
     {
-      "path": ".\\Views\\MasterDetailPage.xaml"
+      "path": ".\\Views\\MasterDetailViewPage.xaml"
     },
     {
-      "path": ".\\Views\\MasterDetailPage.xaml.cs"
+      "path": ".\\Views\\MasterDetailViewPage.xaml.cs"
     },
     {
-      "path": ".\\Views\\MasterDetailDetailPage.xaml"
-    },
-    {
-      "path": ".\\Views\\MasterDetailDetailPage.xaml.cs"
-    },
-    {
-      "path": ".\\Views\\MasterDetailDetailControl.xaml"
-    },
-    {
-      "path": ".\\Views\\MasterDetailDetailControl.xaml.cs"
-    },
-    {
-      "path": ".\\ViewModels\\MasterDetailViewModel.cs"
-    },
-    {
-      "path": ".\\ViewModels\\MasterDetailDetailViewModel.cs"
+      "path": ".\\ViewModels\\MasterDetailViewViewModel.cs"
     }
   ],
   "symbols": {
