--- templates/Pages/MasterDetail.Prism/Views/MasterDetailDetailControl.xaml.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Param_ItemNamespace.Models;
-
-namespace Param_ItemNamespace.Views
-{
-    public sealed partial class MasterDetailDetailControl : UserControl
-    {
-        public SampleOrder MasterMenuItem
-        {
-            get { return GetValue(MasterMenuItemProperty) as SampleOrder; }
-            set { SetValue(MasterMenuItemProperty, value); }
-        }
-
-        public static readonly DependencyProperty MasterMenuItemProperty = DependencyProperty.Register("MasterMenuItem", typeof(SampleOrder), typeof(MasterDetailDetailControl), new PropertyMetadata(null));
-
-        public MasterDetailDetailControl()
-        {
-            InitializeComponent();
-        }
-    }
-}
