--- templates/Pages/MasterDetail.Prism/Views/MasterDetailDetailPage.xaml
@@ -1,22 +1,0 @@
-﻿<Page
-    x:Class="Param_ItemNamespace.Views.MasterDetailDetailPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"    
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:prismMvvm="using:Prism.Windows.Mvvm"
-    prismMvvm:ViewModelLocator.AutoWireViewModel="True"  
-    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
-    xmlns:i="using:Microsoft.Xaml.Interactivity"
-    xmlns:views="using:Param_ItemNamespace.Views"
-    mc:Ignorable="d">
-
-    <Grid
-        x:Name="ContentArea"
-        Background="{ThemeResource SystemControlBackgroundAltHighBrush}">
-
-        <views:MasterDetailDetailControl
-            MasterMenuItem="{x:Bind ViewModel.Item}"
-            Margin="{StaticResource MediumLeftRightMargin}" />
-    </Grid>
-</Page>
