--- templates/Pages/MasterDetail.Prism/Views/MasterDetailPage.xaml
@@ -1,127 +1,0 @@
-﻿<Page
-    x:Class="Param_ItemNamespace.Views.MasterDetailPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:model="using:Param_ItemNamespace.Models"
-    xmlns:views="using:Param_ItemNamespace.Views"
-    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
-    xmlns:i="using:Microsoft.Xaml.Interactivity"
-    mc:Ignorable="d">
-    <Page.Resources>
-        <DataTemplate x:Key="MasterListViewItemTemplate" x:DataType="model:SampleOrder">
-            <Grid Margin="{StaticResource MediumLeftTopRightBottomMargin}" >
-                <StackPanel Orientation="Horizontal" VerticalAlignment="Top" HorizontalAlignment="Left" Grid.Row="1">
-                    <RelativePanel>
-                        <Rectangle Stroke="Gray"
-                             Fill="LightGray"
-                             Width="48"
-                             Height="48"
-                             StrokeThickness="0" 
-                             Name="border"
-                             RelativePanel.AlignVerticalCenterWithPanel="True"/>
-                        <FontIcon
-                            Name="icon"
-                            Tag="{x:Bind HashIdentIcon}"
-                            FontSize="35"
-                            Glyph="{x:Bind SymbolAsChar}" VerticalAlignment="Top" 
-                            RelativePanel.AlignVerticalCenterWith="border"
-                            RelativePanel.AlignHorizontalCenterWith="border"/>
-
-                        <StackPanel RelativePanel.AlignVerticalCenterWithPanel="True" Margin="12,0,0,0" RelativePanel.RightOf="border">
-                            <TextBlock
-                                Name="title"
-                                RelativePanel.RightOf="border"
-                                Text="{x:Bind Company}"
-                                Tag="{x:Bind HashIdentTitle}"
-                                Style="{StaticResource BaseTextBlockStyle}"/>
-                            <TextBlock
-                                Name="subtitle"
-                                RelativePanel.Below="title"
-                                RelativePanel.AlignLeftWith="title"
-                                Text="{x:Bind Status}"
-                                Style="{StaticResource BodyTextBlockStyle}"/>
-                        </StackPanel>
-                    </RelativePanel>
-                </StackPanel>
-            </Grid>
-        </DataTemplate>
-    </Page.Resources>
-
-    <Grid
-        x:Name="ContentArea"
-        Padding="12,0,12,0">
-
-        <i:Interaction.Behaviors>
-            <ic:EventTriggerBehavior EventName="Loaded">
-                <ic:InvokeCommandAction Command="{x:Bind ViewModel.LoadedCommand}" CommandParameter="{Binding ElementName=WindowStates, Path=CurrentState}"/>
-            </ic:EventTriggerBehavior>
-        </i:Interaction.Behaviors>
-
-        <Grid.RowDefinitions>
-            <RowDefinition x:Name="TitleRow" Height="Auto"/>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition x:Name="MasterColumn" Width="296" />
-            <ColumnDefinition x:Name="DetailColumn" Width="*" />
-        </Grid.ColumnDefinitions>
-
-        <TextBlock
-            x:Name="TitlePage"
-            x:Uid="MasterDetailPage_Title"
-            Grid.ColumnSpan="2"
-            Style="{StaticResource PageTitleStyle}"
-            Visibility="Collapsed"/>
-
-        <Grid
-            x:Name="MasterArea"
-            Grid.Row="1"
-            Grid.Column="0">
-            <Grid.RowDefinitions>
-                <RowDefinition x:Name="MasterRow" Height="48"/>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-
-            <TextBlock
-                x:Name="Title"
-                x:Uid="MasterDetailPage_Title"
-                Style="{StaticResource PageTitleStyle}" />
-
-            <!--The SystemControlPageBackgroundChromeLowBrush background represents where you should place your master content.-->
-            <ListView
-                x:Name="MasterListView"
-                Grid.Row="1"
-                ItemsSource="{x:Bind ViewModel.SampleItems, Mode=OneWay}"
-                SelectedItem="{x:Bind ViewModel.Selected, Mode=OneWay}"
-                ItemContainerTransitions="{x:Null}"
-                ItemTemplate="{StaticResource MasterListViewItemTemplate}"
-                IsItemClickEnabled="True">
-                <i:Interaction.Behaviors>
-                    <ic:EventTriggerBehavior EventName="ItemClick">
-                        <ic:InvokeCommandAction Command="{x:Bind ViewModel.ItemClickCommand}" />
-                    </ic:EventTriggerBehavior>
-                </i:Interaction.Behaviors>
-                <ListView.ItemContainerStyle>
-                    <Style TargetType="ListViewItem">
-                        <Setter Property="HorizontalContentAlignment" Value="Stretch" />
-                        <Setter Property="Padding" Value="0"/>
-                    </Style>
-                </ListView.ItemContainerStyle>
-            </ListView>
-        </Grid>
-
-        <Grid
-            x:Name="DetailContentPresenter"
-            Grid.Row="1"
-            Grid.Column="1"
-            Padding="24,0,0,0">
-
-            <views:MasterDetailDetailControl
-                x:Name="DetailControl"
-                MasterMenuItem="{x:Bind ViewModel.Selected, Mode=OneWay}"/>
-        </Grid>
-    </Grid>
-</Page>
