+++ templates/_composition/Prism/Project.SplitView/_postaction.csproj
@@ -1,6 +1,6 @@
 ﻿<!-- Nuget package references -->
 <!--{[{-->
 <PackageReference Include="Microsoft.Toolkit.Uwp.UI.Controls">
-      <Version>1.5.0</Version>
+  <Version>2.0.0</Version>
 </PackageReference>
 <!--}]}-->
