+++ templates/_composition/_shared/Page.MasterDetail/.template.config/template.json
@@ -9,7 +9,7 @@
     "type": "item",
     "wts.type": "composition",
     "wts.version": "1.0.0",
-    "wts.compositionFilter": "$framework == MVVMBasic|MVVMLight|CodeBehind & identity == wts.Page.MasterDetail|wts.Page.MasterDetail.CodeBehind"
+    "wts.compositionFilter": "$framework == MVVMBasic|MVVMLight|CodeBehind|Prism & identity == wts.Page.MasterDetail|wts.Page.MasterDetail.CodeBehind|wts.Page.MasterDetail.Prism"
   },
   "sourceName": "wts.ItemName",
   "preferNameDirectory": true,
